requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.28.1
python-dateutil>=2.8.2

# Confluence API
atlassian-python-api>=3.41.0

# OpenAI
openai>=1.0.0
tiktoken>=0.5.0

# Web Framework
streamlit>=1.28.0

# HTML Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-slugify>=7.0.0
pydantic>=1.10.0
PyYAML>=6.0.0

# Development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
